# tntconcept-rest/tntconcept-openapi.yml
openapi: 3.0.1
info:
  title: tntconcept API
  version: 0.0.1
servers:
- url: http://localhost:8080/
...
paths:
...
  /api/activity:
    get:
      tags:
      - activity
      summary: Gets activities with specified filters
      operationId: get$tntconcept_rest_1
      parameters:
      - name: activityFilterRequest
        in: query
        required: true
        explode: true
        schema:
          $ref: '#/components/schemas/ActivityFilterRequest'
      responses:
        "200":
          description: get$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityResponse'
    put:
      tags:
      - activity
      summary: Updates an existing activity
      operationId: put$tntconcept_rest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
        required: true
      responses:
        "200":
          description: put$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
    post:
      tags:
      - activity
      summary: Creates a new activity
      operationId: post$tntconcept_rest_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
        required: true
      responses:
        "200":
          description: post$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
  /api/activity/summary:
    get:
      tags:
      - activity
      summary: Gets activities summary between two dates
      operationId: summary$tntconcept_rest
      parameters:
      - name: startDate
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: endDate
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: summary$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivitySummaryResponse'
  /api/activity/{id}:
    get:
      tags:
      - activity
      summary: Gets an activity by its id.
      operationId: get$tntconcept_rest_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: get$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
    delete:
      tags:
      - activity
      summary: Deletes an activity by its id.
      operationId: delete$tntconcept_rest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: delete$tntconcept_rest 200 response
  /api/activity/{id}/approve:
    post:
      tags:
      - activity
      summary: Approve an existing activity by id.
      operationId: approve$tntconcept_rest
      requestBody:
        content:
          application/json:
            schema:
              required:
              - locale
              type: integer
              properties:
                locale:
                  $ref: '#/components/schemas/Locale'
              format: int64
        required: true
      responses:
        "200":
          description: approve$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
  /api/activity/{id}/evidence:
    get:
      tags:
      - activity
      summary: Retrieves an activity evidence by the activity id.
      operationId: getEvidenceByActivityId$tntconcept_rest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: getEvidenceByActivityId$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                type: string
...
  /api/vacations:
    get:
      summary: Retrieves holidays within a given charge year.
      operationId: getPrivateHolidaysByChargeYear$tntconcept_rest
      parameters:
      - name: chargeYear
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: getPrivateHolidaysByChargeYear$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayResponse'
    put:
      summary: Updates a holiday period.
      operationId: updatePrivateHolidayPeriod$tntconcept_rest
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateVacationRequest'
              - required:
                - locale
                properties:
                  locale:
                    $ref: '#/components/schemas/Locale'
        required: true
      responses:
        "200":
          description: updatePrivateHolidayPeriod$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreateVacationResponse'
    post:
      summary: Creates a holiday period.
      operationId: createPrivateHolidayPeriod$tntconcept_rest
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateVacationRequest'
              - required:
                - locale
                properties:
                  locale:
                    $ref: '#/components/schemas/Locale'
        required: true
      responses:
        "200":
          description: createPrivateHolidayPeriod$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreateVacationResponse'
  /api/vacations/details:
    get:
      summary: Retrieves details for a holiday within a given charge year.
      operationId: getPrivateHolidayDetails$tntconcept_rest
      parameters:
      - name: chargeYear
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: getPrivateHolidayDetails$tntconcept_rest 200 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacationDetailsResponse'
  /api/vacations/{id}:
    delete:
      summary: Deletes a holiday period by a given ID.
      operationId: deletePrivateHolidayPeriod$tntconcept_rest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: deletePrivateHolidayPeriod$tntconcept_rest 200 response
...
components:
  schemas:
...
    ActivityFilterRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        approvalState:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/ApprovalStateActivityFilter'
        organizationId:
          type: integer
          format: int64
          nullable: true
        projectId:
          type: integer
          format: int64
          nullable: true
        roleId:
          type: integer
          format: int64
          nullable: true
        userId:
          type: integer
          format: int64
          nullable: true
    ActivityRequest:
      required:
      - billable
      - description
      - hasEvidences
      - interval
      - projectRoleId
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        interval:
          $ref: '#/components/schemas/TimeIntervalRequest'
        description:
          maxLength: 2048
          type: string
        billable:
          type: boolean
        projectRoleId:
          type: integer
          format: int64
        hasEvidences:
          type: boolean
        evidence:
          pattern: "^data:[^,]+;base64,.+$"
          type: string
          nullable: true
    ActivityResponse:
      required:
      - approval
      - billable
      - description
      - id
      - interval
      - projectRoleId
      - userId
      type: object
      properties:
        billable:
          type: boolean
        description:
          type: string
        hasEvidences:
          type: boolean
          nullable: true
        id:
          type: integer
          format: int64
        projectRoleId:
          type: integer
          format: int64
        interval:
          $ref: '#/components/schemas/IntervalResponse'
        userId:
          type: integer
          format: int64
        approval:
          $ref: '#/components/schemas/Approval'
    ActivitySummaryResponse:
      required:
      - date
      - worked
      type: object
      properties:
        date:
          type: string
          format: date
        worked:
          type: number
...
    Approval:
      required:
      - canBeApproved
      - state
      type: object
      properties:
        state:
          $ref: '#/components/schemas/ApprovalState'
        canBeApproved:
          type: boolean
        approvedByUserId:
          type: integer
          format: int64
          nullable: true
        approvalDate:
          type: string
          format: date-time
          nullable: true
    ApprovalState:
      type: string
      enum:
      - NA
      - PENDING
      - ACCEPTED
    ApprovalStateActivityFilter:
      type: string
      enum:
      - NA
      - PENDING
      - ACCEPTED
      - ALL
...
    CreateVacationRequest:
      required:
      - endDate
      - startDate
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          maxLength: 1024
          type: string
          nullable: true
    CreateVacationResponse:
      required:
      - chargeYear
      - days
      - endDate
      - startDate
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        days:
          type: integer
          format: int32
        chargeYear:
          type: integer
          format: int32
    HolidayDetailsResponse:
      required:
      - date
      - description
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        date:
          type: string
          format: date
    HolidayResponse:
      required:
      - holidays
      - vacations
      type: object
      properties:
        holidays:
          type: array
          items:
            $ref: '#/components/schemas/HolidayDetailsResponse'
        vacations:
          type: array
          items:
            $ref: '#/components/schemas/VacationResponse'
    IntervalResponse:
      required:
      - duration
      - end
      - start
      - timeUnit
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        duration:
          type: integer
          format: int32
        timeUnit:
          $ref: '#/components/schemas/TimeUnit'
    Locale:
      required:
      - country
      - language
      - variant
      type: object
      properties:
        language:
          type: string
        script:
          type: string
        country:
          type: string
        variant:
          type: string
        extensionKeys:
          type: array
          items:
            type: string
        unicodeLocaleAttributes:
          type: array
          items:
            type: string
        unicodeLocaleKeys:
          type: array
          items:
            type: string
        ISO3Language:
          type: string
        ISO3Country:
          type: string
        displayLanguage:
          type: string
        displayScript:
          type: string
        displayCountry:
          type: string
        displayVariant:
          type: string
        displayName:
          type: string
...
    TimeIntervalRequest:
      required:
      - end
      - start
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
...
    TimeUnit:
      type: string
      enum:
      - MINUTES
      - DAYS
      - NATURAL_DAYS
...
    VacationDetailsResponse:
      required:
      - acceptedVacations
      - correspondingVacations
      - holidaysAgreement
      - remainingVacations
      type: object
      properties:
        holidaysAgreement:
          type: integer
          format: int32
        correspondingVacations:
          type: integer
          format: int32
        acceptedVacations:
          type: integer
          format: int32
        remainingVacations:
          type: integer
          format: int32
    VacationResponse:
      required:
      - chargeYear
      - days
      - endDate
      - startDate
      - state
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        observations:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        state:
          $ref: '#/components/schemas/VacationState'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        days:
          type: array
          items:
            type: string
            format: date
        chargeYear:
          type: string
          format: date
    VacationState:
      type: string
      enum:
      - PENDING
      - ACCEPT
      - REJECT
      - CANCELLED
...